refactor: rollback js module system to CJS
