+++ src/lib/Type.d.ts
